[production/AQM.v7] Add ecFlow option to production branch for AQM (#821)
* add ecflow to prod

* add defn and env file for ecflow

* change template

* fix load modules in scripts

* mv ecf to homedir

* change paths for nco variables

* make complete flag only for rocoto

* bug fix in default fcst_len_cycl

* add missing include

* replace exec with bash for ecflow

* update data_cleanup scripts

* fix bug on aqm_ics

* update aqm_lbcs script

* update ecf suite name

* remove unnecessary lines

* add testing flag

* update sample script

* fix wrong if range

* Fix bug in aqm_manager script

* update aqm_manager script

* move boolify for nco env to job_preamble

+++ scripts/exregional_pre_post_stat.sh
@@ -57,7 +57,7 @@ if [ ${#FCST_LEN_CYCL[@]} -gt 1 ]; then
   cyc_mod=$(( ${cyc} - ${DATE_FIRST_CYCL:8:2} ))
   CYCLE_IDX=$(( ${cyc_mod} / ${INCR_CYCL_FREQ} ))
   FCST_LEN_HRS=${FCST_LEN_CYCL[$CYCLE_IDX]}
-  if [ "${RUN_TASK_RUN_POST}" = "TRUE" ]; then
+  if [ "${WORKFLOW_MANAGER}" = "rocoto" ] && [ "${RUN_TASK_RUN_POST}" = "TRUE" ]; then
     rm -f "${COMIN}/${TN_RUN_POST}_${PDY}${cyc}_task_complete.txt"
   fi
 fi
